+++ lib/src/bones_ui_component.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:html';
 
 import 'package:bones_ui/src/bones_ui_utils.dart';
 import 'package:collection/collection.dart'
@@ -60,7 +59,7 @@ abstract class UIComponent extends UIEventHandler {
 
   bool? get constructing => _constructing;
 
-  UIComponent(UIElement? parent,
+  UIComponent(Object? parent,
       {dynamic componentClass,
       dynamic componentStyle,
       dynamic classes,
@@ -75,12 +74,24 @@ abstract class UIComponent extends UIEventHandler {
       this.id,
       UIComponentGenerator? generator})
       : globalID = ++_globalIDCount,
-        _parent = parent ?? createDivInline(),
         _generator = generator {
-    _resolveParentUIComponent();
+    _resolveParent(parent);
+
     if (construct) {
       _construct(preserveRender, inline, classes, classes2, componentClass,
           style, style2, componentStyle, renderOnConstruction);
+    }
+  }
+
+  void _resolveParent(Object? parent) {
+    _resolveParentUIComponent(parent);
+
+    if (_parent == null) {
+      if (parent is UIElement) {
+        _parent = parent;
+      } else if (parent is UIComponent) {
+        _parent = parent.content;
+      }
     }
   }
 
@@ -105,7 +116,14 @@ abstract class UIComponent extends UIEventHandler {
         _setContent(createContentElement(inline));
       }
 
-      _setParentUIComponent(_getUIComponentByContent(_parent));
+      assert(_content != null);
+
+      if (_parentUIComponent == null && _parent != null) {
+        var uiParent = _getUIComponentByContent(_parent);
+        if (uiParent != null) {
+          _setParentUIComponent(uiParent);
+        }
+      }
 
       registerInUIRoot();
 
@@ -115,9 +133,10 @@ abstract class UIComponent extends UIEventHandler {
       configureStyle(style, style2, componentStyle);
 
       if (clearParent == UIComponentClearParent.onConstruct) {
-        _parent!.nodes.clear();
+        _parent?.nodes.clear();
       }
-      _parent!.append(_content!);
+
+      _parent?.append(_content!);
 
       configure();
 
@@ -144,7 +163,7 @@ abstract class UIComponent extends UIEventHandler {
   static final Expando<UIComponent> _contentsUIComponents =
       Expando<UIComponent>('_content:UIComponent');
 
-  static getContentUIComponent(UIElement content) =>
+  static UIComponent? getContentUIComponent(UIElement content) =>
       _contentsUIComponents[content];
 
   void _setContent(UIElement content) {
@@ -168,7 +187,8 @@ abstract class UIComponent extends UIEventHandler {
 
   /// Called by constructor to register this component in the [UIRoot] tree.
   void registerInUIRoot() {
-    UIRoot.getInstance()!.registerUIComponentInTree(this);
+    var uiRoot = this.uiRoot ?? UIRoot.getInstance();
+    uiRoot?.registerUIComponentInTree(this);
   }
 
   /// Called in the beginning of constructor.
@@ -177,12 +197,12 @@ abstract class UIComponent extends UIEventHandler {
   UIComponent? clone() => null;
 
   /// Sets the [parent] [UIElement].
-  UIElement? setParent(Element parent, {UIComponent? parentUIComponent}) {
+  UIElement? setParent(UIElement parent, {UIComponent? parentUIComponent}) {
     return _setParentImpl(parent, parentUIComponent, true);
   }
 
-  Element? _setParentImpl(
-      Element? parent, UIComponent? parentUIComponent, bool addToParent) {
+  UIElement? _setParentImpl(
+      UIElement? parent, UIComponent? parentUIComponent, bool addToParent) {
     if (parent == null) throw StateError('Null parent');
 
     if (_content != null) {
@@ -190,12 +210,10 @@ abstract class UIComponent extends UIEventHandler {
         if (!identical(_content!.parent, _parent)) {
           _parent!.append(_content!);
         }
-        _resolveParentUIComponent();
+        _resolveParentUIComponent(parentUIComponent ?? parent);
         return _parent;
       } else if (identical(_content!.parent, parent)) {
-        _parentUIComponent = null;
-        _parent = parent;
-        _resolveParentUIComponent();
+        _resolveParentUIComponent(parentUIComponent ?? parent);
         return _parent;
       } else {
         _content!.remove();
@@ -210,8 +228,7 @@ abstract class UIComponent extends UIEventHandler {
       clear();
     }
 
-    _resolveParentUIComponent();
-
+    _resolveParentUIComponent(parentUIComponent ?? parent);
     return _parent;
   }
 
@@ -224,29 +241,47 @@ abstract class UIComponent extends UIEventHandler {
       subUIComponents.expand((e) => [e, ...e.subUIComponents]).toList();
 
   void _setParentUIComponent(UIComponent? uiParent) {
-    _parentUIComponent = uiParent;
-    _uiRoot = null;
+    if (uiParent != null) {
+      _parentUIComponent = uiParent;
+
+      var parentContent = uiParent.content;
+      if (parentContent != null) {
+        var prevParent = _parent;
+
+        if (prevParent == null) {
+          var content = _content;
+          if (identical(content?.parent, parentContent)) {
+            _parent = parentContent;
+          }
+        }
+      }
+
+      _uiRoot = uiParent._uiRoot;
+    } else {
+      _parentUIComponent = null;
+      _uiRoot = null;
+    }
   }
 
   /// The parent [UIComponent].
   UIComponent? get parentUIComponent =>
-      _parentUIComponent ??= _resolveParentUIComponent();
+      _parentUIComponent ??= _resolveParentUIComponent(_parent);
 
-  UIComponent? _resolveParentUIComponent() {
-    UIComponent? foundUIParent;
-
-    var myParent = parent;
-    if (myParent != null) {
-      foundUIParent = getContentUIComponent(myParent);
+  UIComponent? _resolveParentUIComponent(Object? parent) {
+    if (parent is UIComponent) {
+      _setParentUIComponent(parent);
+      return parent;
     }
 
-    foundUIParent ??=
-        _getUIComponentByContent(myParent) ?? _getUIComponentByChild(myParent);
+    if (parent is! UIElement) return null;
 
-    if (foundUIParent != null) {
-      _setParentUIComponent(foundUIParent);
-    }
+    var foundUIParent = getContentUIComponent(parent) ??
+        _getUIComponentByContent(parent) ??
+        _getUIComponentByChild(parent);
 
+    if (foundUIParent == null) return null;
+
+    _setParentUIComponent(foundUIParent);
     return foundUIParent;
   }
 
@@ -261,13 +296,13 @@ abstract class UIComponent extends UIEventHandler {
       return;
     }
 
-    if (elementUIComponent == null && element is Element) {
+    if (elementUIComponent == null && element is UIElement) {
       elementUIComponent = _resolveNodeUIComponent(element);
     }
 
     if (elementUIComponent == null) {
-      if (recursive && element is Element && element.children.isNotEmpty) {
-        if (parentUIComponent == null && parent is Element) {
+      if (recursive && element is UIElement && element.hasUIChildren) {
+        if (parentUIComponent == null && parent is UIElement) {
           parentUIComponent =
               _resolveNodeUIComponent(parent, getUIComponentByChild: true);
         }
@@ -289,7 +324,7 @@ abstract class UIComponent extends UIEventHandler {
       return;
     }
 
-    if (parentUIComponent == null && parent is Element) {
+    if (parentUIComponent == null && parent is UIElement) {
       parentUIComponent =
           _resolveNodeUIComponent(parent, getUIComponentByChild: true);
     }
@@ -297,29 +332,29 @@ abstract class UIComponent extends UIEventHandler {
     if (parentUIComponent != null) {
       elementUIComponent._setParentUIComponent(parentUIComponent);
     } else {
-      elementUIComponent._resolveParentUIComponent();
+      elementUIComponent._resolveParentUIComponent(
+          parent ?? parentUIComponent ?? elementUIComponent._parent);
     }
   }
 
   static UIComponent? _resolveNodeUIComponent(UINode node,
       {bool getUIComponentByChild = false}) {
-    if (node is Element) {
-      var component = UIComponent.getContentUIComponent(node);
-      if (component == null) {
-        var uiRoot = UIRoot.getInstance();
+    if (node is! UIElement) return null;
 
-        if (uiRoot != null) {
-          component =
-              uiRoot.getUIComponentByContent(node, includePurgedEntries: true);
+    var component = UIComponent.getContentUIComponent(node);
+    if (component == null) {
+      var uiRoot = UIRoot.getInstance();
 
-          if (component == null && getUIComponentByChild) {
-            component = uiRoot.getUIComponentByChild(node);
-          }
+      if (uiRoot != null) {
+        component =
+            uiRoot.getUIComponentByContent(node, includePurgedEntries: true);
+
+        if (component == null && getUIComponentByChild) {
+          component = uiRoot.getUIComponentByChild(node);
         }
       }
-      return component;
     }
-    return null;
+    return component;
   }
 
   UIRoot? _uiRoot;
@@ -375,14 +410,14 @@ abstract class UIComponent extends UIEventHandler {
     setID(id);
   }
 
-  void setID(dynamic id) {
+  void setID(Object? id) {
     if (id != null) {
       var idStr = parseString(id, '')!.trim();
-      if (id is String) {
-        id = idStr;
-      }
-
       if (idStr.isNotEmpty) {
+        if (id is String) {
+          id = idStr;
+        }
+
         this.id = id;
         _content!.id = idStr;
         return;
@@ -509,21 +544,22 @@ abstract class UIComponent extends UIEventHandler {
     }
   }
 
-  Element createContentElement(bool inline) {
+  UIElement createContentElement(bool inline) {
     return createDiv(inline);
   }
 
-  Element? get parent => _parent;
+  UIElement? get parent => _parent;
 
-  Element? get content => _content;
+  UIElement? get content => _content;
 
-  List<Element> getContentChildren({FilterElement? filter, bool deep = true}) {
+  List<UIElement> getContentChildren(
+      {FilterElement? filter, bool deep = true}) {
     return _getContentChildrenImpl(
-        _content!.children, <Element>[], deep, filter);
+        _content!.children, <UIElement>[], deep, filter);
   }
 
-  List<Element> _getContentChildrenImpl(
-      List<Element> list, List<Element> dst, bool deep, FilterElement? filter) {
+  List<UIElement> _getContentChildrenImpl(List<UIElement> list,
+      List<UIElement> dst, bool deep, FilterElement? filter) {
     if (list.isEmpty) return dst;
 
     filter ??= (_) => true;
@@ -543,11 +579,11 @@ abstract class UIComponent extends UIEventHandler {
     return dst;
   }
 
-  Element? findInContentChildDeep(FilterElement filter) =>
-      _findInContentChildDeepImpl(_content?.children ?? <Element>[], filter);
+  UIElement? findInContentChildDeep(FilterElement filter) =>
+      _findInContentChildDeepImpl(_content?.children ?? <UIElement>[], filter);
 
-  Element? _findInContentChildDeepImpl(
-      List<Element> list, FilterElement filter) {
+  UIElement? _findInContentChildDeepImpl(
+      List<UIElement> list, FilterElement filter) {
     if (list.isEmpty) return null;
 
     for (var elem in list) {
@@ -562,14 +598,14 @@ abstract class UIComponent extends UIEventHandler {
     return null;
   }
 
-  List<Element> findChildDeep(FilterElement filter) {
-    var list = <Element>[];
+  List<UIElement> findChildDeep(FilterElement filter) {
+    var list = <UIElement>[];
     _findChildDeepImpl(_content!.children, filter, list);
     return list;
   }
 
   void _findChildDeepImpl(
-      List<Element> list, FilterElement filter, List<Element> dst) {
+      List<UIElement> list, FilterElement filter, List<UIElement> dst) {
     if (list.isEmpty) return;
 
     for (var elem in list) {
@@ -587,7 +623,7 @@ abstract class UIComponent extends UIEventHandler {
       _findChildrenDeepImpl(_content!.children, fieldName);
 
   MapEntry<String, Object>? _findChildrenDeepImpl(
-      List<Element> list, String fieldName) {
+      List<UIElement> list, String fieldName) {
     if (list.isEmpty) return null;
 
     for (var elem in list) {
@@ -623,7 +659,7 @@ abstract class UIComponent extends UIEventHandler {
   }
 
   List<MapEntry<String, Object>> _listFieldsEntriesInContentDeepImpl(
-      List<Element> list) {
+      List<UIElement> list) {
     if (list.isEmpty) return <MapEntry<String, Object>>[];
 
     var fieldsEntries = list
@@ -726,7 +762,7 @@ abstract class UIComponent extends UIEventHandler {
       getRenderedElement(
           (elem) =>
               (elem is UIComponent && elem.id == id) ||
-              (elem is Element && elem.id == id),
+              (elem is UIElement && elem.id == id),
           deep);
 
   T? getRenderedUIComponentById<T extends UIComponent>(dynamic id,
@@ -766,7 +802,7 @@ abstract class UIComponent extends UIEventHandler {
       for (var e in _renderedElements!) {
         if (e is UIComponent) {
           e.delete();
-        } else if (e is Element) {
+        } else if (e is UIElement) {
           e.remove();
         }
       }
@@ -927,11 +963,13 @@ abstract class UIComponent extends UIEventHandler {
   static bool get isAnyComponentRendering =>
       UIRoot.getInstance()?.isAnyComponentRendering ?? false;
 
-  static UIComponent? _getUIComponentByContent(Element? content) {
+  static UIComponent? _getUIComponentByContent(UIElement? content) {
+    if (content == null) return null;
     return UIRoot.getInstance()?.getUIComponentByContent(content);
   }
 
-  static UIComponent? _getUIComponentByChild(Element? content) {
+  static UIComponent? _getUIComponentByChild(UIElement? content) {
+    if (content == null) return null;
     return UIRoot.getInstance()?.getUIComponentByChild(content);
   }
 
@@ -965,7 +1003,7 @@ abstract class UIComponent extends UIEventHandler {
     return null;
   }
 
-  UIComponent? findUIComponentByContent(Element? content) {
+  UIComponent? findUIComponentByContent(UIElement? content) {
     if (content == null) return null;
     if (identical(content, _content)) return this;
 
@@ -994,7 +1032,7 @@ abstract class UIComponent extends UIEventHandler {
     return null;
   }
 
-  UIComponent? findUIComponentByChild(Element? child) {
+  UIComponent? findUIComponentByChild(UIElement? child) {
     if (child == null) return null;
     if (identical(child, _content)) return this;
 
@@ -1059,11 +1097,12 @@ abstract class UIComponent extends UIEventHandler {
 
     var content = this.content;
 
-    if (_parent != null) {
+    final parent = _parent;
+    if (parent != null) {
       if (clearParent == UIComponentClearParent.onRender ||
           (clearParent == UIComponentClearParent.onInitialRender &&
               _renderCount == 1)) {
-        var nodes = List<UINode>.from(_parent!.nodes);
+        var nodes = List<UINode>.from(parent.nodes);
 
         var containsContent = false;
         for (var node in nodes) {
@@ -1075,10 +1114,10 @@ abstract class UIComponent extends UIEventHandler {
         }
 
         if (!containsContent) {
-          _parent!.append(content!);
+          parent.append(content!);
         }
-      } else if (!_parent!.nodes.contains(content)) {
-        _parent!.append(content!);
+      } else if (!parent.nodes.contains(content)) {
+        parent.append(content!);
       }
     }
 
@@ -1201,7 +1240,7 @@ abstract class UIComponent extends UIEventHandler {
     for (var e in elements) {
       if (e is UIComponent) {
         e.ensureRendered();
-      } else if (e is Element) {
+      } else if (e is UIElement) {
         var subElements = [];
 
         for (var child in e.children) {
@@ -1341,7 +1380,7 @@ abstract class UIComponent extends UIEventHandler {
   /// Renders the elements of this component.
   ///
   /// Accepted return types:
-  /// - `dart:html` [UINode] and [Element].
+  /// - `dart:html` [UINode] and [UIElement].
   /// - [DIVElement], [DOMNode], [AsDOMElement] and [AsDOMNode].
   /// - [Future].
   /// - [UIAsyncContent].
@@ -1373,7 +1412,7 @@ abstract class UIComponent extends UIEventHandler {
   }
 
   List<dynamic> toContentElements(dynamic rendered,
-      [bool append = false, bool parseAttributes = true]) {
+      {bool append = false, bool parseAttributes = true}) {
     try {
       var list = _toContentElementsImpl(rendered, append);
 
@@ -1391,7 +1430,7 @@ abstract class UIComponent extends UIEventHandler {
 
   static bool isRenderable(dynamic element) {
     if (element == null) return false;
-    return element is Element ||
+    return element is UIElement ||
         element is UIContent ||
         element is UIAsyncContent;
   }
@@ -1411,94 +1450,98 @@ abstract class UIComponent extends UIEventHandler {
 
   static final _listDynamicRuntimeType = <dynamic>[].runtimeType;
 
-  List<dynamic>? toRenderableList(dynamic list) {
-    List<dynamic>? renderableList;
+  List<dynamic>? toRenderableList(Object? list) {
+    if (list == null) return null;
 
-    if (list != null) {
-      if (list is List) {
-        if (list.runtimeType == _listDynamicRuntimeType) {
-          renderableList = list;
-        } else {
-          renderableList = List<dynamic>.from(list);
-        }
-      } else if (list is Iterable) {
-        renderableList = List<dynamic>.from(list);
-      } else if (list is Map) {
-        renderableList = <dynamic>[];
+    List<dynamic> renderableList;
 
-        for (var entry in list.entries) {
-          var key = entry.key;
-          var val = entry.value;
-          if (isRenderable(key)) {
-            renderableList.add(key);
-          }
-          if (isRenderable(val) || isHTMLElement(val)) {
-            renderableList.add(val);
-          }
-        }
+    if (list is List) {
+      if (list.runtimeType == _listDynamicRuntimeType) {
+        renderableList = list;
       } else {
-        renderableList = <dynamic>[list];
+        renderableList = List<dynamic>.from(list);
       }
+    } else if (list is Iterable) {
+      renderableList = List<dynamic>.from(list);
+    } else if (list is Map) {
+      renderableList = <dynamic>[];
+
+      for (var entry in list.entries) {
+        var key = entry.key;
+        var val = entry.value;
+        if (isRenderable(key)) {
+          renderableList.add(key);
+        }
+        if (isRenderable(val) || isHTMLElement(val)) {
+          renderableList.add(val);
+        }
+      }
+    } else {
+      renderableList = <dynamic>[list];
     }
 
     return renderableList;
   }
 
-  List<dynamic> _toContentElementsImpl(dynamic rendered, bool append) {
+  List<dynamic> _toContentElementsImpl(Object? rendered, bool append) {
     var renderableList = toRenderableList(rendered);
 
     var content = this.content!;
 
-    if (renderableList != null) {
-      if (isListOfStrings(renderableList)) {
-        var html = renderableList.join('\n');
-
-        var values = _normalizeRenderListValue(content, html);
-
-        var nodes = (values is List
-                ? values
-                : (values is Iterable ? values.toList() : [values]))
-            .expand((e) => e is List ? e : [e])
-            .map((e) {
-              return _normalizeRenderListValue(content, e);
-            })
-            .cast<UINode>()
-            .toList();
-
-        if (append) {
-          content.nodes.addAll(nodes);
-        } else {
-          content.nodes.clear();
-          content.nodes.addAll(nodes);
-        }
-
-        var renderedList = List<dynamic>.from(content.childNodes);
-        return renderedList;
-      } else {
-        if (isListValuesIdentical(renderableList, content.nodes.toList())) {
-          return List<dynamic>.from(renderableList);
-        }
-
-        for (var value in renderableList) {
-          _removeFromContent(value);
-        }
-
-        var renderedList2 = <dynamic>[];
-
-        var prevElemIndex = -1;
-
-        for (var value in renderableList) {
-          prevElemIndex = _buildRenderList(value, renderedList2, prevElemIndex);
-        }
-
-        return renderedList2;
+    if (renderableList == null || renderableList.isEmpty) {
+      if (!append) {
+        content.nodes.clear();
       }
+
+      return <dynamic>[];
+    }
+
+    if (isListOfStrings(renderableList)) {
+      var html = renderableList.join('\n');
+
+      var values = _normalizeRenderListValue(content, html);
+
+      var nodes = (values is List
+              ? values
+              : (values is Iterable ? values.toList() : [values]))
+          .expand((e) => e is List ? e : [e])
+          .map((e) {
+            return _normalizeRenderListValue(content, e);
+          })
+          .cast<UINode>()
+          .toList();
+
+      if (append) {
+        content.nodes.addAll(nodes);
+      } else {
+        content.nodes.clear();
+        content.nodes.addAll(nodes);
+      }
+
+      var renderedList = List<dynamic>.from(content.childNodes);
+      return renderedList;
     } else {
-      return List<dynamic>.from(content.childNodes);
+      if (isListValuesIdentical(renderableList, content.nodes.toList())) {
+        return List<dynamic>.from(renderableList);
+      }
+
+      for (var value in renderableList) {
+        _removeFromContent(value);
+      }
+
+      var renderedList2 = <dynamic>[];
+
+      var prevElemIndex = -1;
+
+      for (var value in renderableList) {
+        prevElemIndex = _buildRenderList(value, renderedList2, prevElemIndex);
+      }
+
+      return renderedList2;
     }
   }
 
-  dynamic _normalizeRenderListValue(Element? content, value) {
+  dynamic _normalizeRenderListValue(UIElement? content, value) {
     if (value is DOMNode) {
       return value.buildDOM(generator: domGenerator, parent: content);
     } else if (value is AsDOMElement) {
@@ -1661,7 +1704,17 @@ abstract class UIComponent extends UIEventHandler {
 
       _buildRenderList(loadedContent, renderedList, content!.nodes.length - 1);
 
-      content!.nodes.addAll(tail);
+      if (tail.isNotEmpty) {
+        var renderedComponents = renderedList
+            .whereType<UIComponent>()
+            .map((e) => e.content)
+            .toList();
+
+        tail.removeWhere(
+            (e) => renderedList.contains(e) || renderedComponents.contains(e));
+
+        content!.nodes.addAll(tail);
+      }
 
       minRenderedElementsIdx ??= 0;
       _renderedElements!.insertAll(minRenderedElementsIdx, renderedList);
@@ -1736,7 +1789,7 @@ abstract class UIComponent extends UIEventHandler {
   void _removeFromContent(dynamic value) {
     if (value == null) return;
 
-    if (value is Element) {
+    if (value is UIElement) {
       value.remove();
     } else if (value is UIComponent) {
       if (value.isRendered) {
@@ -1756,7 +1809,7 @@ abstract class UIComponent extends UIEventHandler {
     if (list.isEmpty) return;
 
     for (var elem in list) {
-      if (elem is Element) {
+      if (elem is UIElement) {
         _parseNavigate(elem);
         _parseAction(elem);
         _parseEvents(elem);
@@ -1775,7 +1828,7 @@ abstract class UIComponent extends UIEventHandler {
     if (list.isEmpty) return;
 
     for (var elem in list) {
-      if (elem is Element) {
+      if (elem is UIElement) {
         try {
           _parseUiLayout(elem);
         } catch (e) {
@@ -1791,7 +1844,7 @@ abstract class UIComponent extends UIEventHandler {
     }
   }
 
-  void _parseNavigate(Element elem) {
+  void _parseNavigate(UIElement elem) {
     var navigateRoute = getElementAttribute(elem, 'navigate');
 
     if (navigateRoute != null && navigateRoute.isNotEmpty) {
@@ -1799,7 +1852,7 @@ abstract class UIComponent extends UIEventHandler {
     }
   }
 
-  void _parseDataSource(Element content) {
+  void _parseDataSource(UIElement content) {
     var dataSourceCall = getElementAttribute(content, 'data-source');
 
     if (dataSourceCall != null && dataSourceCall.isNotEmpty) {
@@ -2047,7 +2100,7 @@ abstract class UIComponent extends UIEventHandler {
     }
   }
 
-  E? getFieldElement<E extends Element>(String? fieldName) {
+  E? getFieldElement<E extends UIElement>(String? fieldName) {
     var elem =
         findInContentChildDeep((e) => getElementFieldName(e) == fieldName);
     return elem is E ? elem : null;
@@ -2058,53 +2111,33 @@ abstract class UIComponent extends UIEventHandler {
     return findChildrenDeep(fieldName)?.value;
   }
 
-  List<Element> getFieldElements(String? fieldName) =>
+  List<UIElement> getFieldElements(String? fieldName) =>
       findChildDeep((e) => getElementFieldName(e) == fieldName);
 
-  Element? getFieldElementByValue(String? fieldName, String value) =>
+  UIElement? getFieldElementByValue(String? fieldName, String value) =>
       getFieldElements(fieldName)
           .firstWhereOrNull((e) => e.resolveElementValue() == value);
 
   String? getComponentFieldName(Object obj) {
     if (obj is UIField) {
       return obj.fieldName;
-    } else if (obj is Element) {
+    } else if (obj is UIElement) {
       return getElementFieldName(obj);
     } else {
       return null;
     }
   }
 
-  String? getElementFieldName(Element element) {
+  String? getElementFieldName(UIElement element) {
     var ret = _resolveElementField(element);
     return ret?.key;
   }
 
-  MapEntry<String, Object>? _resolveElementField<V>(Element element) {
-    var fieldName = getElementAttributeStr(element, 'field');
-    if (fieldName != null) {
-      fieldName = fieldName.trim();
-      if (fieldName.isNotEmpty) {
-        return MapEntry<String, Object>(fieldName, element);
-      }
-    }
-
-    if (element is InputElement ||
-        element is TextAreaElement ||
-        element is ButtonElement ||
-        element is SelectElement) {
-      fieldName = getElementAttributeStr(element, 'name');
-
-      if (fieldName != null) {
-        fieldName = fieldName.trim();
-        if (fieldName.isNotEmpty) {
-          return MapEntry<String, Object>(fieldName, element);
-        }
-      }
-    }
+  MapEntry<String, Object>? _resolveElementField<V>(UIElement element) {
+    var fieldName = element.resolveFieldName();
+    if (fieldName != null) return fieldName;
 
     var component = _getUIComponentByContent(element);
-
     if (component != null) {
       if (component is UIField) {
         var field = component as UIField;
@@ -2116,7 +2149,7 @@ abstract class UIComponent extends UIEventHandler {
     return null;
   }
 
-  Map<String, Element> getFieldsElementsMap(
+  Map<String, UIElement> getFieldsElementsMap(
       {List<String>? fields, List<String>? ignoreFields}) {
     ignoreFields ??= [];
 
@@ -2124,7 +2157,7 @@ abstract class UIComponent extends UIEventHandler {
 
     var fieldsElements = getFieldsElements();
 
-    var map = <String, Element>{};
+    var map = <String, UIElement>{};
 
     for (var elem in fieldsElements) {
       var fieldName = getElementFieldName(elem)!;
@@ -2152,24 +2185,29 @@ abstract class UIComponent extends UIEventHandler {
   }
 
   /// Alias to [content.querySelector].
-  E? querySelector<E extends Element>(String? selectors) {
+  E? querySelector<E extends UIElement>(String? selectors) {
     if (selectors == null || selectors.isEmpty) return null;
-    var elem = content?.querySelector(selectors);
-    return elem is E ? elem : null;
+    if (E == UIElement) {
+      var elem = content?.querySelector(selectors);
+      return elem is E ? elem : null;
+    } else {
+      return content?.querySelectorAll(selectors).whereType<E>().firstOrNull;
+    }
   }
 
   /// Alias to [content.querySelector].
-  E? selectElement<E extends Element>(String? selectors) =>
+  E? selectElement<E extends UIElement>(String? selectors) =>
       querySelector<E>(selectors);
 
   /// Alias to [content.querySelectorAll].
-  List<T> querySelectorAll<T extends Element>(String? selectors) {
+  List<T> querySelectorAll<T extends UIElement>(String? selectors) {
     if (selectors == null || selectors.isEmpty) return <T>[];
-    return content?.querySelectorAll(selectors) ?? <T>[];
+    return content?.querySelectorAll(selectors).whereType<T>().toList() ??
+        <T>[];
   }
 
   /// Alias to [content.querySelectorAll].
-  List<T> selectElements<T extends Element>(String? selectors) =>
+  List<T> selectElements<T extends UIElement>(String? selectors) =>
       querySelectorAll(selectors);
 
   bool clearContent() {
@@ -2188,10 +2226,10 @@ abstract class UIComponent extends UIEventHandler {
     return true;
   }
 
-  List<Element> getFieldsElements() => _getContentChildrenImpl(
+  List<UIElement> getFieldsElements() => _getContentChildrenImpl(
       content!.children, [], true, (e) => getElementFieldName(e) != null);
 
-  String? parseChildElementValue(Element? childElement,
+  String? parseChildElementValue(UIElement? childElement,
           {UIComponent? childUiComponent, bool allowTextAsValue = true}) =>
       childElement?.resolveElementValue(
           parentUIComponent: this,
@@ -2284,11 +2322,7 @@ abstract class UIComponent extends UIEventHandler {
     _renderedFieldsValues ??= {};
     _renderedFieldsValues![fieldName] = valueStr;
 
-    if (fieldElem is InputElement) {
-      fieldElem.value = valueStr;
-    } else {
-      fieldElem.text = valueStr;
-    }
+    fieldElem.setValue(valueStr);
   }
 
   void updateRenderedFieldValue(String fieldName) {
@@ -2301,7 +2335,7 @@ abstract class UIComponent extends UIEventHandler {
     renderedFieldsValues[fieldName] = value;
   }
 
-  void updateRenderedFieldElementValue(Element fieldElem) {
+  void updateRenderedFieldElementValue(UIElement fieldElem) {
     var fieldName = getElementFieldName(fieldElem);
     if (fieldName == null) return;
 
@@ -2332,7 +2366,7 @@ abstract class UIComponent extends UIEventHandler {
       var val = parseChildElementValue(fieldComponent.content,
           childUiComponent: fieldComponent);
       return val as V? ?? def;
-    } else if (fieldComponent is Element) {
+    } else if (fieldComponent is UIElement) {
       var val = parseChildElementValue(fieldComponent);
       return val as V? ?? def;
     } else {
@@ -2553,12 +2587,11 @@ abstract class UIComponent extends UIEventHandler {
 
     var component = getFieldComponent(fieldName);
 
-    if (component is Element) {
+    if (component is UIElement) {
       component.focus();
       return true;
     } else if (component is UIComponent) {
-      var input = findInContentChildDeep(
-          (e) => e is InputElement || e is TextAreaElement);
+      var input = findInContentChildDeep((e) => e.isTextInput);
       if (input != null) {
         input.focus();
         return true;
@@ -2624,7 +2657,7 @@ abstract class UIComponent extends UIEventHandler {
     return count;
   }
 
-  void _parseAction(Element elem) {
+  void _parseAction(UIElement elem) {
     var actionValue = getElementAttribute(elem, 'action');
 
     if (actionValue != null && actionValue.isNotEmpty) {
@@ -2636,7 +2669,7 @@ abstract class UIComponent extends UIEventHandler {
     UIConsole.log('action: $action');
   }
 
-  void _parseUiLayout(Element elem) {
+  void _parseUiLayout(UIElement elem) {
     var uiLayout = getElementAttribute(elem, 'uiLayout');
 
     if (uiLayout != null) {
@@ -2644,12 +2677,12 @@ abstract class UIComponent extends UIEventHandler {
     }
   }
 
-  void _parseEvents(Element elem) {
+  void _parseEvents(UIElement elem) {
     _parseOnEventKeyPress(elem);
     _parseOnEventClick(elem);
   }
 
-  void _parseOnEventKeyPress(Element elem) {
+  void _parseOnEventKeyPress(UIElement elem) {
     var keypress = getElementAttribute(elem, 'onEventKeyPress');
 
     if (keypress != null && keypress.isNotEmpty) {
@@ -2671,7 +2704,7 @@ abstract class UIComponent extends UIEventHandler {
     }
   }
 
-  void _parseOnEventClick(Element elem) {
+  void _parseOnEventClick(UIElement elem) {
     var click = getElementAttribute(elem, 'onEventClick');
 
     if (click != null && click.isNotEmpty) {