+++ lib/src/bones_ui_navigator.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:html';
 
 import 'package:collection/collection.dart';
 import 'package:swiss_knife/swiss_knife.dart';
@@ -20,9 +19,9 @@ class UINavigator {
   }
 
   UINavigator._() {
-    window.onHashChange.listen((e) => _onChangeRoute(e as HashChangeEvent));
+    navigationOnChangeRoute(_onChangeRoute);
 
-    var href = window.location.href;
+    var href = navigationURL();
     var url = Uri.parse(href);
 
     var routeFragment = _parseRouteFragment(url);
@@ -38,7 +37,7 @@ class UINavigator {
   }
 
   /// Returns [true] if this device is online.
-  static bool get isOnline => window.navigator.onLine ?? false;
+  static bool get isOnline => navigationIsOnline();
 
   /// Returns [true] if this device is off-line.
   static bool get isOffline => !isOnline;
@@ -48,17 +47,19 @@ class UINavigator {
   /// See: [https://developer.mozilla.org/en-US/docs/Web/Security/Secure_Contexts]
   static bool get isSecureContext {
     try {
-      return window.isSecureContext ?? false;
+      return navigationIsSecureContext();
     } catch (e, s) {
       logger.error('Error calling `window.isSecureContext`', e, s);
       return false;
     }
   }
 
-  void _onChangeRoute(HashChangeEvent event) {
-    var uri = Uri.parse(event.newUrl!);
-    UIConsole.log(
-        'UINavigator._onChangeRoute: new: $uri > previous: ${event.oldUrl}');
+  void _onChangeRoute(String? oldURL, String? newUrl) {
+    if (newUrl == null) return;
+
+    var uri = Uri.parse(newUrl);
+    UIConsole.log('UINavigator._onChangeRoute: new: $uri > previous: $oldURL');
+
     _navigateToFromURL(uri);
   }
 
@@ -415,7 +416,7 @@ class UINavigator {
 
     if (routeQueryString.isNotEmpty) fragment += '?$routeQueryString';
 
-    var locationUrl = window.location.href;
+    var locationUrl = navigationURL();
     var locationUrl2 = locationUrl.contains('#')
         ? locationUrl.replaceFirst(RegExp(r'#.*'), fragment)
         : '$locationUrl$fragment';
@@ -426,7 +427,7 @@ class UINavigator {
     }
 
     if (!fromURL) {
-      window.history.pushState({}, routeTitle, locationUrl2);
+      navigationHistoryPush(routeTitle, locationUrl2);
     }
 
     clearDetachedNavigables();
@@ -497,7 +498,7 @@ class UINavigator {
   List<UIElement> selectNavigables([UIElement? element]) {
     return element != null
         ? element.querySelectorAll(_navigableComponentSelector)
-        : document.querySelectorAll(_navigableComponentSelector);
+        : documentQuerySelectorAll(_navigableComponentSelector);
   }
 
   /// Find in [element] tree nodes with attribute `navigate`.
@@ -768,7 +769,7 @@ abstract class UINavigableComponent extends UIComponent {
 
     this.findRoutes = findRoutes;
 
-    UIConsole.log('_normalizeRoutes: $_routes -> $routesOk');
+    // UIConsole.log('_normalizeRoutes: $_routes -> $routesOk');
 
     _routes = routesOk;
   }
@@ -946,7 +947,7 @@ abstract class UINavigableContent extends UINavigableComponent {
     List allRendered = [];
 
     if (topMargin > 0) {
-      var divTopMargin = Element.div();
+      var divTopMargin = UIElement.div();
       divTopMargin.style.width = '100%';
       divTopMargin.style.height = '${topMargin}px';