+++ CHANGELOG.md
@@ -1,8 +1,16 @@
-##
+## 2.4.0
 
 - New `bones_ui_web.dart`:
-  - New  `UIElement` and `UINode`.
+  - Avoid imports to `dart:html` in preparation to use of `dart:web` or other alternatives.
+  - New typedef `UINode` and `UIElement`.
 - Replace usage of `Element` to `UIElement` and `Node` to `UINode`.
+- `UIComponent`:
+  - Constructor:
+    - parameter `parent` now accepts `Element` or `UIComponent`.
+  - `toContentElements`: changed to named parameters.
+
+- intl_messages: ^2.2.3
+- dom_tools: ^2.2.0
 
 ## 2.3.3